Attempted to change the way items are taken out of the inventory after save
The way items were being taken out of the inventory in postAutoSave() did not work with stacks of items. It was only taking things out of a single slot while stacks of items might be sorted into multiple slots.

The new method I am trying is supposed to look through every slot in the inventory and take out the correct number of items from any stack of the correct item. Unfortunately the way I am comparing items does not seem to work with stackable items. It will work correctly with single items like tools, but comparing a stack of blocks to another stack of blocks does not seem to return true. I think if I could compare the items correctly this would work the way it should.