feat: add inventory screen titles (#45)
This adds a (translated) title to the inventory screen.

In itself this doesn't do much yet and might seem unnecessary. However, especially when using item blocks that have an inventory (chests, assembly tables, etc.), it easily becomes confusing, which inventory is the own and which is the inventory of the block being interacted with. Hence, I added this for container inventory as well.

If the container parent prefab does not have a `DisplayNameComponent`, the prefab name will be shown, for instance `CoreAdvancedAssets:chest`.

Co-authored-by: Tobias Nett <skaldarnar@googlemail.com>