fix: retrieve character entity within binding
In #45 the character entity was cached outside of the binding
evaluation. However, it in multiplayer games it can happen that it is
`null`, and therefore opening container screens on a connected client
fails.

This is fixed by retrieving the character entity with each binding
evaluation via a common helper method `getPredictedInteractionTarget`.